Extract BubbleKebabMenu + generalize Bubble actions prop
- New BubbleKebabMenu component owns the kebab icon, dropdown panel,
  outside-click-to-close, and the visible/menu-open visibility math.
  Rendered exactly the same way it used to be inline in Bubble.
- Bubble swaps enableActions+onDeleteThread for a generic
  actions: Array<{label, action, color?}> prop. Kebab renders only
  when actions has items AND the bubble is not a tombstone/inaccessible
  placeholder (preserving the existing isPlaceholder short-circuit).
- Feed.js builds the prior Delete-thread item via the new actions prop.
